+++ packages/form_builder_extra_fields/lib/src/fields/form_builder_searchable_dropdown.dart
@@ -192,7 +192,7 @@ class FormBuilderSearchableDropdown<T> extends FormBuilderField<T> {
   final PositionCallback? positionCallback;
 
   /// Creates field for selecting value(s) from a searchable list
-  FormBuilderSearchableDropdown( {
+  FormBuilderSearchableDropdown({
     Key? key,
     //From Super
     required String name,