+++ packages/form_builder_extra_fields/lib/src/fields/form_builder_typeahead.dart
@@ -261,7 +261,7 @@ class FormBuilderTypeAhead<T> extends FormBuilderField<T> {
   final ScrollController? scrollController;
 
   /// Creates text field that auto-completes user input from a list of items
-  FormBuilderTypeAhead( {
+  FormBuilderTypeAhead({
     Key? key,
     //From Super
     required String name,